requirements.txt
requests==2.32.3
PyYAML==6.0.2
playwright==1.47.0
